feat: update render methods to use float for mouse coordinates for improved precision

+++ src/main/java/dev/anvilcraft/resource/ageratum/client/feat/markdown/component/MDHorizontalRuleComponent.java
@@ -19,7 +19,7 @@ public class MDHorizontalRuleComponent extends MDComponent {
      * 在当前行中间绘制一条水平线。
      */
     @Override
-    public void render(GuiGraphics guiGraphics, Minecraft minecraft, int maxX, int maxY, int mouseX, int mouseY) {
+    public void render(GuiGraphics guiGraphics, Minecraft minecraft, int maxX, int maxY, float mouseX, float mouseY) {
         int y = minecraft.font.lineHeight / 2;
         guiGraphics.hLine(0, Math.max(0, maxX - 1), y, 0x88000000);
     }
